requirements.txt
# Core ML/AI Framework
tensorflow>=2.16.0
numpy>=1.21.0,<2.0.0
scipy>=1.7.0

# Computer Vision
opencv-python>=4.5.0
Pillow>=8.3.0

# Data Science
pandas>=1.3.0
scikit-learn>=1.0.0
matplotlib>=3.5.0
seaborn>=0.11.0

# Image Processing & Augmentation
tensorflow-keras-preprocessing>=1.1.2

# Utilities
tqdm>=4.62.0
argparse

# Development & Testing
pytest>=6.2.0
pytest-cov>=2.12.0
black>=21.0.0
flake8>=3.9.0

# API Development (Optional)
flask>=2.0.0
fastapi>=0.70.0
uvicorn>=0.15.0

# Jupyter Notebook Support
jupyter>=1.0.0
ipykernel>=6.0.0
notebook>=6.4.0

# File I/O
pathlib
glob2
shutil

# Optional: For deployment
gunicorn>=20.1.0
docker

# Model Optimization (Optional)
tensorflow-model-optimization>=0.7.0
tensorboard>=2.8.0